Добавил все Controller, Service, DAO interface, осталось реализовать DAO

+++ src/main/java/com/palex/practice/service/UserService.java
@@ -12,7 +12,7 @@ public interface UserService {
     /**
      *  Отобразить объект по Id.
      */
-    void getById(Integer id);
+    void getById(Long id);
 
     /**
      * Изменить объект.
